requirements/requirements_juan.txt

absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.86
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dm-env==1.6
dm-tree==0.1.9
docker-pycreds==0.4.0
docstring-parser==0.16
envpool==0.8.4
etils==1.12.0
executing==2.2.0
farama-notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.18.0
fire==0.7.0
flax==0.8.4
fqdn==1.5.1
fsspec==2024.3.1
gitdb==4.0.12
gitpython==3.1.44
gym==0.23.1
gym-notices==0.0.8
gymnasium==1.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-auto-schema==0.0.7
hydra-core==1.3.2
hydra-zen==0.13.0
idna==3.10
importlib-resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.25
jax-cuda12-pjrt==0.4.25
jax-cuda12-plugin==0.4.25
jaxlib==0.4.25
jaxtyping==0.2.38
jedi==0.19.2
jinja2==3.1.6
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.8.61
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.1.3
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
omegaconf==2.3.0
opt-einsum==3.4.0
optax==0.2.2
optree==0.14.1
orbax-checkpoint==0.5.18
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pytest==8.3.5
pytest-datadir==1.6.1
pytest-regressions==2.7.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pyyaml==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.2
send2trash==1.8.3
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tensorstore==0.1.72
termcolor==3.1.0
terminado==0.18.1
tinycss2==1.4.0
toolz==1.0.0
# torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.1.5
types-protobuf==5.29.1.20250208
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
uri-template==1.3.0
urllib3==1.26.20
wadler-lindig==0.1.3
wandb==0.19.7
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xtils @ git+https://github.com/jessefarebro/xtils@cf368978cadb9ea5d5befc73a25b5ef004ac0fa7
zipp==3.21.0
